Crate trillium_rustls

This crate provides rustls trait implementations for trillium client (RustlsConnector) and server (RustlsAcceptor).

Re-exports

pub use async_rustls;

pub use rustls;

Structs

RustlsAcceptor

trillium Acceptor for Rustls

RustlsConfig

Client configuration for RustlsConnector

RustlsConnector

this struct provides rustls a trillium client connector implementation