Struct trillium_rustls::RustlsTransport

pub struct RustlsTransport<T>(/* private fields */);

Transport for the rustls connector

This may represent either an encrypted tls connection or a plaintext connection, depending on the request schema